requirements.txt
aiosqlite==0.21.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
idna==3.10
mysql-connector-python==9.3.0
parameterized==0.9.0
pillow==11.2.1
pycodestyle==2.5.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
